+++ src/meshes/mesh_io.jl
@@ -6,7 +6,7 @@
 #! format: noindent
 
 # Save current mesh with some context information as an HDF5 file.
-function save_mesh_file(mesh::Union{TreeMesh, P4estMesh}, output_directory,
+function save_mesh_file(mesh::Union{TreeMesh, P4estMesh, T8codeMesh}, output_directory,
                         timestep = 0)
     save_mesh_file(mesh, output_directory, timestep, mpi_parallel(mesh))
 end
@@ -222,7 +222,14 @@ end
 
 function save_mesh_file(mesh::T8codeFVMesh, output_directory,
                         timestep = 0)
-    "empty_mesh_file"
+    joinpath(output_directory, "dummy_mesh.h5")
+end
+
+# TODO: Implement this function as soon as there is support for this in `t8code`.
+function save_mesh_file(mesh::T8codeMesh, output_directory, timestep, mpi_parallel)
+    error("Mesh file output not supported yet for `T8codeMesh`.")
+
+    return joinpath(output_directory, "dummy_mesh.h5")
 end
 
 """
@@ -261,36 +268,32 @@ function load_mesh_serial(mesh_file::AbstractString; n_cells_max, RealT)
         size = Tuple(size_)
 
         # TODO: `@eval` is evil
-        # A temporary workaround to evaluate the code that defines the domain mapping in a local scope.
-        # This prevents errors when multiple restart elixirs are executed in one session, where one
-        # defines `mapping` as a variable, while the other defines it as a function.
         #
         # This should be replaced with something more robust and secure,
         # see https://github.com/trixi-framework/Trixi.jl/issues/541).
-        expr = Meta.parse(mapping_as_string)
-        if expr.head == :toplevel
-            expr.head = :block
-        end
-
         if ndims == 1
-            mapping = @eval function (xi)
-                $expr
+            mapping = eval(Meta.parse("""function (xi)
+                $mapping_as_string
                 mapping(xi)
             end
+            """))
         elseif ndims == 2
-            mapping = @eval function (xi, eta)
-                $expr
+            mapping = eval(Meta.parse("""function (xi, eta)
+                $mapping_as_string
                 mapping(xi, eta)
             end
+            """))
         else # ndims == 3
-            mapping = @eval function (xi, eta, zeta)
-                $expr
+            mapping = eval(Meta.parse("""function (xi, eta, zeta)
+                $mapping_as_string
                 mapping(xi, eta, zeta)
             end
+            """))
         end
 
         mesh = StructuredMesh(size, mapping; RealT = RealT, unsaved_changes = false,
                               mapping_as_string = mapping_as_string)
+        mesh.current_filename = mesh_file
     elseif mesh_type == "UnstructuredMesh2D"
         mesh_filename, periodicity_ = h5open(mesh_file, "r") do file
             return read(attributes(file)["mesh_filename"]),